ref: create a `metrics` module
