+++ src/tools/metrics/mod.rs
@@ -1,0 +1,4 @@
+//! Metrics types and utilites.
+
+pub mod recorder;
+pub mod tables;
