+++ src/tools/mod.rs
@@ -1,7 +1,7 @@
 //! Utilities for network testing.
 
 pub mod message_filter;
-pub mod simple_metrics;
+pub mod metrics;
 pub mod synthetic_node;
 
 use std::time::Duration;
